*리스트 정렬(최신순, 평점순) =>문제점 : 정렬 제대로 안됨 *검색했을때(엔터or클릭) 스크롤을 최상단으로 올라가게 만들 예정
-성현

+++ recipemaster/src/main/java/com/recipe/domain/Recipe.java
@@ -18,7 +18,7 @@ public class Recipe {
 	private Date recipeDate;
 	private int hits;
 	private int regiStatus;
-	private int gradePoint;
+	private double gradePoint;
 	private User user;
 	private Date likeDate;
 	private int likeUser;
@@ -112,10 +112,10 @@ public class Recipe {
 	public void setRegiStatus(int regiStatus) {
 		this.regiStatus = regiStatus;
 	}
-	public int getGradePoint() {
+	public double getGradePoint() {
 		return gradePoint;
 	}
-	public void setGradePoint(int gradePoint) {
+	public void setGradePoint(double gradePoint) {
 		this.gradePoint = gradePoint;
 	}
 	public User getUser() {
